+++ src/extra/translator/__init__.py
@@ -1,5 +1,2 @@
 # -*- coding: utf-8 -*-
-from __future__ import absolute_import
-from __future__ import unicode_literals
-from . import translator
-from . import wx_ui as gui
+from .translator import TranslatorController